agents/requirements.txt
google-adk>=0.5.0
openpyxl>=3.1.0
xlrd>=2.0.1
pandas>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-Levenshtein>=0.21.0
